Move 3 hazard unittest
Move tests:

1) climada/hazard/test/test_base/test_write_read_pass  (4.5 sec)
2)climada/hazard/test/test_base/test_raster_to_vector_pass (3.5 sec)
3)climada/hazard/test/test_base/test_reproject_raster_pass (3.6 sec)

Destination:

1) climada/test/test_hazard/test_write_read_pass
2) climada/test/test_hazard/test_raster_to_vector_pass
3) climada/test/test_hazard/test_reproject_raster_pass
