+++ climada/test/test_hazard.py
@@ -22,6 +22,7 @@ Test Hazard base class.
 import unittest
 import numpy as np
 import datetime as dt
+from pathlib import Path
 from scipy import sparse
 
 from climada import CONFIG
@@ -31,8 +32,13 @@ from climada.hazard.centroids import Centroids
 from climada.hazard.storm_europe import StormEurope
 from climada.util.constants import (HAZ_DEMO_FL, WS_DEMO_NC)
 from climada.util.api_client import Client
+from climada.util import coordinates as u_coord
+import climada.hazard.test as hazard_test
 
 DATA_DIR = CONFIG.test_data.dir()
+
+# Hazard test file from Git repository. Fraction is 1. Format: matlab.
+HAZ_TEST_MAT :Path = Path(hazard_test.__file__).parent.joinpath('data', 'atl_prob_no_name.mat')
 
 class TestCentroids(unittest.TestCase):
     """Test centroids functionalities"""
@@ -246,9 +252,102 @@ class TestTcTracks(unittest.TestCase):
         self.assertIn('The following generated synthetic tracks moved beyond '
                       'the range of [-70, 70] degrees latitude', cm.output[1])
 
+class TestBase(unittest.TestCase):
+    """ Test methods to create Hazard objects from hdf5, matlab mat
+        and raster."""
+
+    def test_write_read_pass(self):
+        """Read a hazard mat file correctly."""
+        
+        file_name = str(DATA_DIR.joinpath('test_haz.h5'))
+        # Read demo matlab file
+        hazard = Hazard.from_mat(HAZ_TEST_MAT)
+        hazard.event_name = list(map(str, hazard.event_name))
+        for todense_flag in [False, True]:
+            if todense_flag:
+                hazard.write_hdf5(file_name, todense=todense_flag)
+            else:
+                hazard.write_hdf5(file_name)
+
+            haz_read = Hazard.from_hdf5(file_name)
+
+            self.assertEqual(str(hazard.tag.file_name), haz_read.tag.file_name)
+            self.assertIsInstance(haz_read.tag.file_name, str)
+            self.assertEqual(hazard.tag.haz_type, haz_read.tag.haz_type)
+            self.assertIsInstance(haz_read.tag.haz_type, str)
+            self.assertEqual(hazard.tag.description, haz_read.tag.description)
+            self.assertIsInstance(haz_read.tag.description, str)
+            self.assertEqual(hazard.units, haz_read.units)
+            self.assertIsInstance(haz_read.units, str)
+            self.assertTrue(np.array_equal(hazard.centroids.coord, haz_read.centroids.coord))
+            self.assertTrue(u_coord.equal_crs(hazard.centroids.crs, haz_read.centroids.crs))
+            self.assertTrue(np.array_equal(hazard.event_id, haz_read.event_id))
+            self.assertTrue(np.array_equal(hazard.frequency, haz_read.frequency))
+            self.assertEqual(hazard.frequency_unit, haz_read.frequency_unit)
+            self.assertIsInstance(haz_read.frequency_unit, str)
+            self.assertTrue(np.array_equal(hazard.event_name, haz_read.event_name))
+            self.assertIsInstance(haz_read.event_name, list)
+            self.assertIsInstance(haz_read.event_name[0], str)
+            self.assertTrue(np.array_equal(hazard.date, haz_read.date))
+            self.assertTrue(np.array_equal(hazard.orig, haz_read.orig))
+            self.assertTrue(np.array_equal(hazard.intensity.toarray(),
+                                           haz_read.intensity.toarray()))
+            self.assertIsInstance(haz_read.intensity, sparse.csr_matrix)
+            self.assertTrue(np.array_equal(hazard.fraction.toarray(), haz_read.fraction.toarray()))
+            self.assertIsInstance(haz_read.fraction, sparse.csr_matrix)
+
+    def test_raster_to_vector_pass(self):
+        """Test raster_to_vector method"""
+        
+        haz_fl = Hazard.from_raster([HAZ_DEMO_FL], haz_type='FL')
+        haz_fl.check()
+        meta_orig = haz_fl.centroids.meta
+        inten_orig = haz_fl.intensity
+        fract_orig = haz_fl.fraction
+
+        haz_fl.raster_to_vector()
+
+        self.assertEqual(haz_fl.centroids.meta, dict())
+        self.assertAlmostEqual(haz_fl.centroids.lat.min(),
+                            meta_orig['transform'][5]
+                            + meta_orig['height'] * meta_orig['transform'][4]
+                            - meta_orig['transform'][4] / 2)
+        self.assertAlmostEqual(haz_fl.centroids.lat.max(),
+                            meta_orig['transform'][5] + meta_orig['transform'][4] / 2)
+        self.assertAlmostEqual(haz_fl.centroids.lon.max(),
+                            meta_orig['transform'][2]
+                            + meta_orig['width'] * meta_orig['transform'][0]
+                            - meta_orig['transform'][0] / 2)
+        self.assertAlmostEqual(haz_fl.centroids.lon.min(),
+                            meta_orig['transform'][2] + meta_orig['transform'][0] / 2)
+        self.assertTrue(u_coord.equal_crs(haz_fl.centroids.crs, meta_orig['crs']))
+        self.assertTrue(np.allclose(haz_fl.intensity.data, inten_orig.data))
+        self.assertTrue(np.allclose(haz_fl.fraction.data, fract_orig.data))
+
+    def test_reproject_raster_pass(self):
+        """Test reproject_raster reference."""
+        
+        haz_fl = Hazard.from_raster([HAZ_DEMO_FL])
+        haz_fl.check()
+
+        haz_fl.reproject_raster(dst_crs='epsg:2202')
+
+        self.assertEqual(haz_fl.intensity.shape, (1, 1046408))
+        self.assertIsInstance(haz_fl.intensity, sparse.csr_matrix)
+        self.assertIsInstance(haz_fl.fraction, sparse.csr_matrix)
+        self.assertEqual(haz_fl.fraction.shape, (1, 1046408))
+        self.assertTrue(u_coord.equal_crs(haz_fl.centroids.meta['crs'], 'epsg:2202'))
+        self.assertEqual(haz_fl.centroids.meta['width'], 968)
+        self.assertEqual(haz_fl.centroids.meta['height'], 1081)
+        self.assertEqual(haz_fl.fraction.min(), 0)
+        self.assertEqual(haz_fl.fraction.max(), 1)
+        self.assertEqual(haz_fl.intensity.min(), -9999)
+        self.assertTrue(haz_fl.intensity.max() < 4.7)
+
 # Execute Tests
 if __name__ == "__main__":
     TESTS = unittest.TestLoader().loadTestsFromTestCase(TestCentroids)
     TESTS.addTest(unittest.TestLoader().loadTestsFromTestCase(TestStormEurope))
     TESTS.addTest(unittest.TestLoader().loadTestsFromTestCase(TestTcTracks))
+    TESTS.addTest(unittest.TestLoader().loadTestsFromTestCase(TestBase))
     unittest.TextTestRunner(verbosity=2).run(TESTS)
